+++ src/pages/createBill/addExpenseStep/index.tsx
@@ -1,35 +1,23 @@
-import { useLayoutEffect, useMemo, useRef, useState } from 'react';
-import { FormProvider, useFieldArray, useForm } from 'react-hook-form';
-import { zodResolver } from '@hookform/resolvers/zod';
+import { useLayoutEffect, useRef } from 'react';
+import { FormProvider } from 'react-hook-form';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { Close } from '@/assets/svgs/icon';
 import Header from '@/common/components/Header';
 import { BaseFunnelStepComponentProps } from '@/common/types/useFunnel.type';
-import { Group } from '@/common/types/group.type';
-import {
-  Expense,
-  ExpenseFormSchema,
-} from '@/pages/createBill/types/expense.type';
 import expense from '@/service/apis/expense';
-import group from '@/service/apis/group';
 import FormCard from '@/pages/createBill/addExpenseStep/components/FormCard';
+import useAddExpenseFormArray from '@/pages/createBill/hooks/useAddExpenseFormArray';
 import getTotalExpense from '@/pages/createBill/utils/getTotalExpense';
 import { BillContext } from '@/pages/createBill/types/billContext.type';
 import * as S from './index.styles';
-
-const defaultValues: Omit<Expense, 'id'> = {
-  amount: 0,
-  content: '',
-  date: new Date(),
-  memberExpenses: [],
-};
 
 interface AddExpenseStepProps
   extends BaseFunnelStepComponentProps<BillContext> {}
 
 function AddExpenseStep({ moveToNextStep }: AddExpenseStepProps) {
   const lastFormCardRef = useRef<HTMLDivElement | null>(null);
-  const [groupInfo, setGroupInfo] = useState<Group | null>(null);
+  const { groupInfo, formMethods, defaultFormValue, fieldArrayReturns } =
+    useAddExpenseFormArray();
   const queryClient = useQueryClient();
   const mutation = useMutation({
     mutationFn: expense.create,
@@ -40,52 +28,13 @@ function AddExpenseStep({ moveToNextStep }: AddExpenseStepProps) {
       moveToNextStep?.();
     },
   });
-  const formMethods = useForm({
-    resolver: zodResolver(ExpenseFormSchema),
-    mode: 'onChange', // 폼들의 필수 입력값이 모두 입력되었을 때 '다음' 버튼을 활성화시키기 위함
-    defaultValues: async () => {
-      // TODO : groupToken을 받아오는 로직 추가
-      const groupData = await group.get('groupToken');
-      setGroupInfo(groupData);
-      return {
-        expenses: [
-          {
-            ...defaultValues,
-            memberExpenses: groupData.members.map((member) => ({
-              memberId: member.id,
-              name: member.name,
-              amount: 0,
-            })),
-          },
-        ],
-      };
-    },
-  });
-  // TODO : 리팩토링 필요할듯...
-  const defaultFormValue = useMemo(() => {
-    if (!groupInfo) {
-      return defaultValues;
-    }
-    return {
-      ...defaultValues,
-      memberExpenses: groupInfo.members.map((member) => ({
-        memberId: member.id,
-        name: member.name,
-        amount: 0,
-      })),
-    };
-  }, [groupInfo]);
-  const { fields, append, remove } = useFieldArray({
-    control: formMethods.control,
-    name: 'expenses',
-  });
   useLayoutEffect(() => {
     // form의 개수가 변경되면 (추가, 삭제) 마지막 form으로 스크롤 이동
     lastFormCardRef.current?.scrollIntoView({
       behavior: 'smooth',
       block: 'start',
     });
-  }, [fields.length]);
+  }, [fieldArrayReturns.fields.length]);
 
   const { handleSubmit, formState, watch } = formMethods;
   const allFormsValid = formState.isValid;
@@ -93,11 +42,10 @@ function AddExpenseStep({ moveToNextStep }: AddExpenseStepProps) {
 
   const handleAddExpense = () => {
     // 기본 focus 기능을 사용하지 않고 새로운 폼 추가
-    append(defaultFormValue, { shouldFocus: false });
+    fieldArrayReturns.append(defaultFormValue, { shouldFocus: false });
   };
-
   const handleDeleteExpense = (index: number) => {
-    remove(index);
+    fieldArrayReturns.remove(index);
   };
 
   if (!groupInfo) {
@@ -119,10 +67,14 @@ function AddExpenseStep({ moveToNextStep }: AddExpenseStepProps) {
         </S.TopMessage>
       </S.TopWrapper>
       <S.BillFormList>
-        {fields.map((field, index) => (
+        {fieldArrayReturns.fields.map((field, index) => (
           <FormCard
             key={field.id}
-            ref={index === fields.length - 1 ? lastFormCardRef : null}
+            ref={
+              index === fieldArrayReturns.fields.length - 1
+                ? lastFormCardRef
+                : null
+            }
             index={index}
             onDelete={handleDeleteExpense}
           />
